added pretty-print capabilities
Added the ability to output a formatted XML document tree, with
indentation and newlines, when outputting children and siblings,
respectively.

As is the case with all 'pretty-print' capabilities of other libraries,
the document tree output is not the same. The resultant text if parsed
will have a tree that will contain more/less XmlText nodes than the
original tree.