Preserve the single/double quotes in attributes between parsing and writing.
This changes the interface of XmlGrammar from "createAttribute(XmlName, String)"
to "createAttribute(XmlName, String, XmlAttributeType)" and thus break clients
that override this method. Otherwise the change should be backward compatible.

The attributes in the XmlBuilder take a new optional argument that specifies
if single or double quotes should be generated (double quotes by default).